+++ stylua-vscode/src/stylua.ts
@@ -45,6 +45,15 @@ export function formatCode(
       args.push("--range-end");
       args.push(endPos.toString());
     }
+
+    const configPath = vscode.workspace
+      .getConfiguration("stylua")
+      .get<string>("configPath");
+    if (configPath && configPath.trim() !== "") {
+      args.push("--config-path");
+      args.push(configPath);
+    }
+
     if (
       vscode.workspace.getConfiguration("stylua").get("searchParentDirectories")
     ) {